Feat: 메인페이지 기본 레이아웃 구성
