+++ components/layout/footer.tsx
@@ -1,0 +1,3 @@
+export default function Footer() {
+  return <footer>footer</footer>;
+}
