+++ components/layout/header.tsx
@@ -1,0 +1,7 @@
+export default function Header() {
+  return (
+    <header className="w-full border-b px-2 py-4">
+      <h1 className="font-bold">wwg</h1>
+    </header>
+  );
+}
