+++ next.config.ts
@@ -2,6 +2,14 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   /* config options here */
+  images: {
+    remotePatterns: [
+      // {
+      //   protocol: "http",
+      //   hostname: "localhost",
+      // },
+    ],
+  },
 };
 
 export default nextConfig;
